Use elm-sortable-table
I don't really need sorting, but this library handles a lot of other
stuff related to displaying tables, so should prevent me from writing
a lot of boilerplate.

I just tacked a sortable table on the bottom of the view
that displays model.table for now.
I will delete the original view stuff and put input fields in the
sortable table next.