Put light background behind corrected SG and ABV
This makes it the table easier to understand when these columns are
empty because you see an empty gray box instead of just blank space.
It's especially confusing without this because the last 3 table headers
look smashed together. That would be the correct problem to fix if I
wanted this app to look good, but I just need this to be easy to use.