Add button to clear table
This button deletes everything in the table.
Which reveals a cosmetic bug:
After clearing, the first row will show the values written in the
first row before it was cleared, despite the values in the model
being reset to Nothing.
This is because I use Html.Attributes.defaultValue instead of
Html.Attributes.value to set the text on the input elements.

This allowed me to exploit the memory of html input elements to make
typing feel natural with very little effort. If I want to use
Html.Attributes.value instead, I will have to change the model to hold
strings for each input field because numbers can change when converted
from a string to a number and then back
eg:
    "1.0" |> String.toFloat |> toString
    returns: "1"
which makes it very difficult to type values like 1.05.