Store lastRow separately from rest of rows
The table should always have at least one row, and the last row of the
table should:
    1. always be blank
    and
    2. add a new row if it's modified.

So, in the spirit of Richard Feldman's "Making Impossible States
Impossible", I'm storing the table in two separate parts of the
model:
{ ...
  , table : List Row
  , lastRow : Row
  ...
}

Now I never have to worry about whether the table is empty or not,
and it's easier to handle the special logic associated with the last
row.

This commit introduces one small user visible change:
when the defaultCalibration is modified, it will update the lastRow's
calibration.