Remove default calibration field
Default calibration is now automatically updated whenever the
penultimate row's calibration changes. The effect of this is
that whenever a new entry is added with a new calibration, that
calibration is copied to subsequent rows until it is changed again.

Compared to the separate default calibration field, this is more
intuitive, requires strictly less typing, and sacrifices no
functionality.