Improve keyboard navigation
Each row has 3 input fields and a delete button.

old tab order:
Pressing tab would go through each input on a row, then move to its
delete button, and then move to the next row.

It is extremely rare that i want to enter data for a row and then
immediately delete it, so i moved the delete buttons out of this flow.

New tab order:
Now, pressing tab goes through the input fields on a row, then goes to
the next row. At the end of the table, it will move to the first row's
delete button, and then tab will move to the next row's delete button.

I find this much more convenient, both for deleting and editing rows.

I would pretty much never use tabindex normally because it is really
easy to mess it up and make your website annoying for people using
accessibility software, however, this is app is mostly for my own use
and I think it really does improve keyboard navigation (at the cost of
violating the principle of least surprise).