Fix inconsistent calibration change behavior
When you changed the calibration of the very last row, a new row would be
created, with the calibration set to the old default calibration.
If you then modified the calibration of the second to last row (which used to
be the last one) the current last row would be updated to the expected
value.
A very very tiny and weird bug. One of the very few times I've had to
spend more than a minute debugging a problem in elm code.